+++ CHANGELOG.md
@@ -5,10 +5,10 @@
 - added misc module
   - stash highlight on `alt+f`
 - added trade module
-  - support for poe.app, poemap.live, poetrade, official trade
+  - support for poe.app, poemap.live, poe.trade, official trade
   - view incoming/ outgoing trades
   - invite/ trade player
-  - highlight item in stash
+  - highlight items in stash
 - fixed a first time launching error occuring if the GGG API could not be reached
 
 ## 1.0.0 (2020-06-10)